BAH-4721 | Refactor. Rename Medication Input Control As MedicationRequest
