+++ apps/clinical/src/components/forms/medicationRequest/SelectedMedicationItem.tsx
