+++ apps/clinical/src/components/forms/medicationRequest/__tests__/SelectedMedicationItem.test.tsx
