+++ apps/clinical/src/components/forms/medicationRequest/styles/SelectedMedicationItem.module.scss
